LayoutTests/fast/dom/wrapper-classes-expected.txt

This tests wrapper class names for JavaScript.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Basics

PASS jsWrapperClass(document.createAttribute()) is 'Attr'
PASS jsWrapperClass(document.createAttribute().__proto__) is 'AttrPrototype'
PASS jsWrapperClass(document.createAttribute().constructor) is 'Function'
PASS jsWrapperClass(document.createComment()) is 'Comment'
PASS jsWrapperClass(document.createComment().__proto__) is 'CommentPrototype'
PASS jsWrapperClass(document.createComment().constructor) is 'Function'
PASS jsWrapperClass(document.createDocumentFragment()) is 'DocumentFragment'
PASS jsWrapperClass(document.createDocumentFragment().__proto__) is 'DocumentFragmentPrototype'
PASS jsWrapperClass(document.createDocumentFragment().constructor) is 'Function'
PASS jsWrapperClass(document.implementation) is 'DOMImplementation'
PASS jsWrapperClass(document.implementation.__proto__) is 'DOMImplementationPrototype'
PASS jsWrapperClass(document.implementation.constructor) is 'Function'
PASS jsWrapperClass(root.attributes) is 'NamedNodeMap'
PASS jsWrapperClass(root.attributes.__proto__) is 'NamedNodeMapPrototype'
PASS jsWrapperClass(root.attributes.constructor) is 'Function'
PASS jsWrapperClass(document.createNodeIterator(root, 0, null, false)) is 'NodeIterator'
PASS jsWrapperClass(document.createNodeIterator(root, 0, null, false).__proto__) is 'NodeIteratorPrototype'
PASS jsWrapperClass(document.createNodeIterator(root, 0, null, false).constructor) is 'Function'
PASS jsWrapperClass(document.getElementsByTagName('div')) is 'HTMLCollection'
PASS jsWrapperClass(document.getElementsByTagName('div').__proto__) is 'HTMLCollectionPrototype'
PASS jsWrapperClass(document.getElementsByTagName('div').constructor) is 'Function'
PASS jsWrapperClass(document.createRange()) is 'Range'
PASS jsWrapperClass(document.createRange().__proto__) is 'RangePrototype'
PASS jsWrapperClass(document.createRange().constructor) is 'Function'
PASS jsWrapperClass(document.createTextNode()) is 'Text'
PASS jsWrapperClass(document.createTextNode().__proto__) is 'TextPrototype'
PASS jsWrapperClass(document.createTextNode().constructor) is 'Function'
PASS jsWrapperClass(document.createTreeWalker(root, 0, null, false)) is 'TreeWalker'
PASS jsWrapperClass(document.createTreeWalker(root, 0, null, false).__proto__) is 'TreeWalkerPrototype'
PASS jsWrapperClass(document.createTreeWalker(root, 0, null, false).constructor) is 'Function'

XML

PASS jsWrapperClass(xmlDocument) is 'XMLDocument'
PASS jsWrapperClass(xmlDocument.__proto__) is 'XMLDocumentPrototype'
PASS jsWrapperClass(xmlDocument.constructor) is 'Function'
PASS jsWrapperClass(xmlDocument.createCDATASection()) is 'CDATASection'
PASS jsWrapperClass(xmlDocument.createCDATASection().__proto__) is 'CDATASectionPrototype'
PASS jsWrapperClass(xmlDocument.createCDATASection().constructor) is 'Function'
PASS jsWrapperClass(xmlDocument.createElementNS('x', 'x', 'x')) is 'Element'
PASS jsWrapperClass(xmlDocument.createElementNS('x', 'x', 'x').__proto__) is 'ElementPrototype'
PASS jsWrapperClass(xmlDocument.createElementNS('x', 'x', 'x').constructor) is 'Function'
PASS jsWrapperClass(xmlDocument.createProcessingInstruction()) is 'ProcessingInstruction'
PASS jsWrapperClass(xmlDocument.createProcessingInstruction().__proto__) is 'ProcessingInstructionPrototype'
PASS jsWrapperClass(xmlDocument.createProcessingInstruction().constructor) is 'Function'

Events

PASS jsWrapperClass(document.createEvent('Event')) is 'Event'
PASS jsWrapperClass(document.createEvent('Event').__proto__) is 'EventPrototype'
PASS jsWrapperClass(document.createEvent('Event').constructor) is 'Function'
PASS jsWrapperClass(document.createEvent('KeyboardEvent')) is 'KeyboardEvent'
PASS jsWrapperClass(document.createEvent('KeyboardEvent').__proto__) is 'KeyboardEventPrototype'
PASS jsWrapperClass(document.createEvent('KeyboardEvent').constructor) is 'Function'
PASS jsWrapperClass(document.createEvent('MouseEvent')) is 'MouseEvent'
PASS jsWrapperClass(document.createEvent('MouseEvent').__proto__) is 'MouseEventPrototype'
PASS jsWrapperClass(document.createEvent('MouseEvent').constructor) is 'Function'
PASS jsWrapperClass(document.createEvent('MutationEvent')) is 'MutationEvent'
PASS jsWrapperClass(document.createEvent('MutationEvent').__proto__) is 'MutationEventPrototype'
PASS jsWrapperClass(document.createEvent('MutationEvent').constructor) is 'Function'
PASS jsWrapperClass(document.createEvent('OverflowEvent')) is 'OverflowEvent'
PASS jsWrapperClass(document.createEvent('OverflowEvent').__proto__) is 'OverflowEventPrototype'
PASS jsWrapperClass(document.createEvent('OverflowEvent').constructor) is 'Function'
PASS jsWrapperClass(document.createEvent('TextEvent')) is 'TextEvent'
PASS jsWrapperClass(document.createEvent('TextEvent').__proto__) is 'TextEventPrototype'
PASS jsWrapperClass(document.createEvent('TextEvent').constructor) is 'Function'
PASS jsWrapperClass(document.createEvent('UIEvent')) is 'UIEvent'
PASS jsWrapperClass(document.createEvent('UIEvent').__proto__) is 'UIEventPrototype'
PASS jsWrapperClass(document.createEvent('UIEvent').constructor) is 'Function'
PASS jsWrapperClass(document.createEvent('WheelEvent')) is 'WheelEvent'
PASS jsWrapperClass(document.createEvent('WheelEvent').__proto__) is 'WheelEventPrototype'
PASS jsWrapperClass(document.createEvent('WheelEvent').constructor) is 'Function'

CSS DOM

PASS jsWrapperClass(document.styleSheets) is 'StyleSheetList'
PASS jsWrapperClass(document.styleSheets.__proto__) is 'StyleSheetListPrototype'
PASS jsWrapperClass(document.styleSheets.constructor) is 'Function'
PASS jsWrapperClass(stylesheet) is 'CSSStyleSheet'
PASS jsWrapperClass(stylesheet.__proto__) is 'CSSStyleSheetPrototype'
PASS jsWrapperClass(stylesheet.constructor) is 'Function'
PASS jsWrapperClass(stylesheet.cssRules) is 'CSSRuleList'
PASS jsWrapperClass(stylesheet.cssRules.__proto__) is 'CSSRuleListPrototype'
PASS jsWrapperClass(stylesheet.cssRules.constructor) is 'Function'
PASS jsWrapperClass(stylesheet.cssRules.item(0)) is 'CSSStyleRule'
PASS jsWrapperClass(stylesheet.cssRules.item(0).__proto__) is 'CSSStyleRulePrototype'
PASS jsWrapperClass(stylesheet.cssRules.item(0).constructor) is 'Function'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style) is 'CSSStyleDeclaration'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.__proto__) is 'CSSStyleDeclarationPrototype'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.constructor) is 'Function'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color')) is 'CSSPrimitiveValue'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color').__proto__) is 'CSSPrimitiveValuePrototype'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color').constructor) is 'Function'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color').getRGBColorValue()) is 'RGBColor'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color').getRGBColorValue().__proto__) is 'RGBColorPrototype'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('color').getRGBColorValue().constructor) is 'Function'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('clip').getRectValue()) is 'Rect'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('clip').getRectValue().__proto__) is 'RectPrototype'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('clip').getRectValue().constructor) is 'Function'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content')) is 'CSSValueList'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content').__proto__) is 'CSSValueListPrototype'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content').constructor) is 'Function'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content').item(0).getCounterValue()) is 'Counter'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content').item(0).getCounterValue().__proto__) is 'CounterPrototype'
PASS jsWrapperClass(stylesheet.cssRules.item(0).style.getPropertyCSSValue('content').item(0).getCounterValue().constructor) is 'Function'
PASS jsWrapperClass(stylesheet.media) is 'MediaList'
PASS jsWrapperClass(stylesheet.media.__proto__) is 'MediaListPrototype'
PASS jsWrapperClass(stylesheet.media.constructor) is 'Function'

XPath

PASS jsWrapperClass(document.createExpression('//*', document.createNSResolver(document))) is 'XPathExpression'
PASS jsWrapperClass(document.createExpression('//*', document.createNSResolver(document)).__proto__) is 'XPathExpressionPrototype'
PASS jsWrapperClass(document.createExpression('//*', document.createNSResolver(document)).constructor) is 'Function'
PASS jsWrapperClass(document.evaluate('//*', document, document.createNSResolver(document), 0, null)) is 'XPathResult'
PASS jsWrapperClass(document.evaluate('//*', document, document.createNSResolver(document), 0, null).__proto__) is 'XPathResultPrototype'
PASS jsWrapperClass(document.evaluate('//*', document, document.createNSResolver(document), 0, null).constructor) is 'Function'

Other

PASS jsWrapperClass(window) is 'Window'
PASS jsWrapperClass(window.__proto__) is 'WindowPrototype'
PASS jsWrapperClass(window.constructor) is 'Function'

HTML DOM

PASS jsWrapperClass(root.children) is 'HTMLCollection'
PASS jsWrapperClass(root.children.__proto__) is 'HTMLCollectionPrototype'
PASS jsWrapperClass(root.children.constructor) is 'Function'
PASS jsWrapperClass(document) is 'HTMLDocument'
PASS jsWrapperClass(document.__proto__) is 'HTMLDocumentPrototype'
PASS jsWrapperClass(document.constructor) is 'Function'
PASS jsWrapperClass(document.createElement('select').options) is 'HTMLOptionsCollection'
PASS jsWrapperClass(document.createElement('select').options.__proto__) is 'HTMLOptionsCollectionPrototype'
PASS jsWrapperClass(document.createElement('select').options.constructor) is 'Function'

HTML Elements

PASS tagJSWrapperClass('a') is 'HTMLAnchorElement'
PASS tagJSWrapperPrototypeClass('a') is 'HTMLAnchorElementPrototype'
PASS tagJSWrapperConstructorClass('a') is 'Function'
PASS tagJSWrapperClass('abbr') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('abbr') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('abbr') is 'Function'
PASS tagJSWrapperClass('acronym') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('acronym') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('acronym') is 'Function'
PASS tagJSWrapperClass('address') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('address') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('address') is 'Function'
PASS tagJSWrapperClass('applet') is 'HTMLAppletElement'
PASS tagJSWrapperPrototypeClass('applet') is 'HTMLAppletElementPrototype'
PASS tagJSWrapperConstructorClass('applet') is 'Function'
PASS tagJSWrapperClass('area') is 'HTMLAreaElement'
PASS tagJSWrapperPrototypeClass('area') is 'HTMLAreaElementPrototype'
PASS tagJSWrapperConstructorClass('area') is 'Function'
PASS tagJSWrapperClass('article') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('article') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('article') is 'Function'
PASS tagJSWrapperClass('aside') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('aside') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('aside') is 'Function'
PASS tagJSWrapperClass('b') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('b') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('b') is 'Function'
PASS tagJSWrapperClass('base') is 'HTMLBaseElement'
PASS tagJSWrapperPrototypeClass('base') is 'HTMLBaseElementPrototype'
PASS tagJSWrapperConstructorClass('base') is 'Function'
PASS tagJSWrapperClass('bdo') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('bdo') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('bdo') is 'Function'
PASS tagJSWrapperClass('big') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('big') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('big') is 'Function'
PASS tagJSWrapperClass('blockquote') is 'HTMLQuoteElement'
PASS tagJSWrapperPrototypeClass('blockquote') is 'HTMLQuoteElementPrototype'
PASS tagJSWrapperConstructorClass('blockquote') is 'Function'
PASS tagJSWrapperClass('body') is 'HTMLBodyElement'
PASS tagJSWrapperPrototypeClass('body') is 'HTMLBodyElementPrototype'
PASS tagJSWrapperConstructorClass('body') is 'Function'
PASS tagJSWrapperClass('br') is 'HTMLBRElement'
PASS tagJSWrapperPrototypeClass('br') is 'HTMLBRElementPrototype'
PASS tagJSWrapperConstructorClass('br') is 'Function'
PASS tagJSWrapperClass('button') is 'HTMLButtonElement'
PASS tagJSWrapperPrototypeClass('button') is 'HTMLButtonElementPrototype'
PASS tagJSWrapperConstructorClass('button') is 'Function'
PASS tagJSWrapperClass('canvas') is 'HTMLCanvasElement'
PASS tagJSWrapperPrototypeClass('canvas') is 'HTMLCanvasElementPrototype'
PASS tagJSWrapperConstructorClass('canvas') is 'Function'
PASS tagJSWrapperClass('caption') is 'HTMLTableCaptionElement'
PASS tagJSWrapperPrototypeClass('caption') is 'HTMLTableCaptionElementPrototype'
PASS tagJSWrapperConstructorClass('caption') is 'Function'
PASS tagJSWrapperClass('center') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('center') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('center') is 'Function'
PASS tagJSWrapperClass('cite') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('cite') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('cite') is 'Function'
PASS tagJSWrapperClass('code') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('code') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('code') is 'Function'
PASS tagJSWrapperClass('col') is 'HTMLTableColElement'
PASS tagJSWrapperPrototypeClass('col') is 'HTMLTableColElementPrototype'
PASS tagJSWrapperConstructorClass('col') is 'Function'
PASS tagJSWrapperClass('colgroup') is 'HTMLTableColElement'
PASS tagJSWrapperPrototypeClass('colgroup') is 'HTMLTableColElementPrototype'
PASS tagJSWrapperConstructorClass('colgroup') is 'Function'
PASS tagJSWrapperClass('dd') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('dd') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('dd') is 'Function'
PASS tagJSWrapperClass('del') is 'HTMLModElement'
PASS tagJSWrapperPrototypeClass('del') is 'HTMLModElementPrototype'
PASS tagJSWrapperConstructorClass('del') is 'Function'
PASS tagJSWrapperClass('dfn') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('dfn') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('dfn') is 'Function'
PASS tagJSWrapperClass('dir') is 'HTMLDirectoryElement'
PASS tagJSWrapperPrototypeClass('dir') is 'HTMLDirectoryElementPrototype'
PASS tagJSWrapperConstructorClass('dir') is 'Function'
PASS tagJSWrapperClass('div') is 'HTMLDivElement'
PASS tagJSWrapperPrototypeClass('div') is 'HTMLDivElementPrototype'
PASS tagJSWrapperConstructorClass('div') is 'Function'
PASS tagJSWrapperClass('dl') is 'HTMLDListElement'
PASS tagJSWrapperPrototypeClass('dl') is 'HTMLDListElementPrototype'
PASS tagJSWrapperConstructorClass('dl') is 'Function'
PASS tagJSWrapperClass('dt') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('dt') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('dt') is 'Function'
PASS tagJSWrapperClass('em') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('em') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('em') is 'Function'
PASS tagJSWrapperClass('embed') is 'HTMLEmbedElement'
PASS tagJSWrapperPrototypeClass('embed') is 'HTMLEmbedElementPrototype'
PASS tagJSWrapperConstructorClass('embed') is 'Function'
PASS tagJSWrapperClass('fieldset') is 'HTMLFieldSetElement'
PASS tagJSWrapperPrototypeClass('fieldset') is 'HTMLFieldSetElementPrototype'
PASS tagJSWrapperConstructorClass('fieldset') is 'Function'
PASS tagJSWrapperClass('font') is 'HTMLFontElement'
PASS tagJSWrapperPrototypeClass('font') is 'HTMLFontElementPrototype'
PASS tagJSWrapperConstructorClass('font') is 'Function'
PASS tagJSWrapperClass('footer') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('footer') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('footer') is 'Function'
PASS tagJSWrapperClass('form') is 'HTMLFormElement'
PASS tagJSWrapperPrototypeClass('form') is 'HTMLFormElementPrototype'
PASS tagJSWrapperConstructorClass('form') is 'Function'
PASS tagJSWrapperClass('frame') is 'HTMLFrameElement'
PASS tagJSWrapperPrototypeClass('frame') is 'HTMLFrameElementPrototype'
PASS tagJSWrapperConstructorClass('frame') is 'Function'
PASS tagJSWrapperClass('frameset') is 'HTMLFrameSetElement'
PASS tagJSWrapperPrototypeClass('frameset') is 'HTMLFrameSetElementPrototype'
PASS tagJSWrapperConstructorClass('frameset') is 'Function'
PASS tagJSWrapperClass('head') is 'HTMLHeadElement'
PASS tagJSWrapperPrototypeClass('head') is 'HTMLHeadElementPrototype'
PASS tagJSWrapperConstructorClass('head') is 'Function'
PASS tagJSWrapperClass('h1') is 'HTMLHeadingElement'
PASS tagJSWrapperPrototypeClass('h1') is 'HTMLHeadingElementPrototype'
PASS tagJSWrapperConstructorClass('h1') is 'Function'
PASS tagJSWrapperClass('h2') is 'HTMLHeadingElement'
PASS tagJSWrapperPrototypeClass('h2') is 'HTMLHeadingElementPrototype'
PASS tagJSWrapperConstructorClass('h2') is 'Function'
PASS tagJSWrapperClass('h3') is 'HTMLHeadingElement'
PASS tagJSWrapperPrototypeClass('h3') is 'HTMLHeadingElementPrototype'
PASS tagJSWrapperConstructorClass('h3') is 'Function'
PASS tagJSWrapperClass('h4') is 'HTMLHeadingElement'
PASS tagJSWrapperPrototypeClass('h4') is 'HTMLHeadingElementPrototype'
PASS tagJSWrapperConstructorClass('h4') is 'Function'
PASS tagJSWrapperClass('h5') is 'HTMLHeadingElement'
PASS tagJSWrapperPrototypeClass('h5') is 'HTMLHeadingElementPrototype'
PASS tagJSWrapperConstructorClass('h5') is 'Function'
PASS tagJSWrapperClass('h6') is 'HTMLHeadingElement'
PASS tagJSWrapperPrototypeClass('h6') is 'HTMLHeadingElementPrototype'
PASS tagJSWrapperConstructorClass('h6') is 'Function'
PASS tagJSWrapperClass('head') is 'HTMLHeadElement'
PASS tagJSWrapperPrototypeClass('head') is 'HTMLHeadElementPrototype'
PASS tagJSWrapperConstructorClass('head') is 'Function'
PASS tagJSWrapperClass('header') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('header') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('header') is 'Function'
PASS tagJSWrapperClass('hgroup') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('hgroup') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('hgroup') is 'Function'
PASS tagJSWrapperClass('hr') is 'HTMLHRElement'
PASS tagJSWrapperPrototypeClass('hr') is 'HTMLHRElementPrototype'
PASS tagJSWrapperConstructorClass('hr') is 'Function'
PASS tagJSWrapperClass('html') is 'HTMLHtmlElement'
PASS tagJSWrapperPrototypeClass('html') is 'HTMLHtmlElementPrototype'
PASS tagJSWrapperConstructorClass('html') is 'Function'
PASS tagJSWrapperClass('i') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('i') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('i') is 'Function'
PASS tagJSWrapperClass('iframe') is 'HTMLIFrameElement'
PASS tagJSWrapperPrototypeClass('iframe') is 'HTMLIFrameElementPrototype'
PASS tagJSWrapperConstructorClass('iframe') is 'Function'
PASS tagJSWrapperClass('image') is 'HTMLUnknownElement'
PASS tagJSWrapperPrototypeClass('image') is 'HTMLUnknownElementPrototype'
PASS tagJSWrapperConstructorClass('image') is 'Function'
PASS tagJSWrapperClass('img') is 'HTMLImageElement'
PASS tagJSWrapperPrototypeClass('img') is 'HTMLImageElementPrototype'
PASS tagJSWrapperConstructorClass('img') is 'Function'
PASS tagJSWrapperClass('input') is 'HTMLInputElement'
PASS tagJSWrapperPrototypeClass('input') is 'HTMLInputElementPrototype'
PASS tagJSWrapperConstructorClass('input') is 'Function'
PASS tagJSWrapperClass('ins') is 'HTMLModElement'
PASS tagJSWrapperPrototypeClass('ins') is 'HTMLModElementPrototype'
PASS tagJSWrapperConstructorClass('ins') is 'Function'
PASS tagJSWrapperClass('kbd') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('kbd') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('kbd') is 'Function'
PASS tagJSWrapperClass('keygen') is 'HTMLKeygenElement'
PASS tagJSWrapperPrototypeClass('keygen') is 'HTMLKeygenElementPrototype'
PASS tagJSWrapperConstructorClass('keygen') is 'Function'
PASS tagJSWrapperClass('label') is 'HTMLLabelElement'
PASS tagJSWrapperPrototypeClass('label') is 'HTMLLabelElementPrototype'
PASS tagJSWrapperConstructorClass('label') is 'Function'
PASS tagJSWrapperClass('layer') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('layer') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('layer') is 'Function'
PASS tagJSWrapperClass('legend') is 'HTMLLegendElement'
PASS tagJSWrapperPrototypeClass('legend') is 'HTMLLegendElementPrototype'
PASS tagJSWrapperConstructorClass('legend') is 'Function'
PASS tagJSWrapperClass('li') is 'HTMLLIElement'
PASS tagJSWrapperPrototypeClass('li') is 'HTMLLIElementPrototype'
PASS tagJSWrapperConstructorClass('li') is 'Function'
PASS tagJSWrapperClass('link') is 'HTMLLinkElement'
PASS tagJSWrapperPrototypeClass('link') is 'HTMLLinkElementPrototype'
PASS tagJSWrapperConstructorClass('link') is 'Function'
PASS tagJSWrapperClass('listing') is 'HTMLPreElement'
PASS tagJSWrapperPrototypeClass('listing') is 'HTMLPreElementPrototype'
PASS tagJSWrapperConstructorClass('listing') is 'Function'
PASS tagJSWrapperClass('main') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('main') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('main') is 'Function'
PASS tagJSWrapperClass('map') is 'HTMLMapElement'
PASS tagJSWrapperPrototypeClass('map') is 'HTMLMapElementPrototype'
PASS tagJSWrapperConstructorClass('map') is 'Function'
PASS tagJSWrapperClass('marquee') is 'HTMLMarqueeElement'
PASS tagJSWrapperPrototypeClass('marquee') is 'HTMLMarqueeElementPrototype'
PASS tagJSWrapperConstructorClass('marquee') is 'Function'
PASS tagJSWrapperClass('menu') is 'HTMLMenuElement'
PASS tagJSWrapperPrototypeClass('menu') is 'HTMLMenuElementPrototype'
PASS tagJSWrapperConstructorClass('menu') is 'Function'
PASS tagJSWrapperClass('meta') is 'HTMLMetaElement'
PASS tagJSWrapperPrototypeClass('meta') is 'HTMLMetaElementPrototype'
PASS tagJSWrapperConstructorClass('meta') is 'Function'
PASS tagJSWrapperClass('nav') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('nav') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('nav') is 'Function'
PASS tagJSWrapperClass('nobr') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('nobr') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('nobr') is 'Function'
PASS tagJSWrapperClass('noembed') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('noembed') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('noembed') is 'Function'
PASS tagJSWrapperClass('noframes') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('noframes') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('noframes') is 'Function'
PASS tagJSWrapperClass('nolayer') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('nolayer') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('nolayer') is 'Function'
PASS tagJSWrapperClass('noscript') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('noscript') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('noscript') is 'Function'
PASS tagJSWrapperClass('object') is 'HTMLObjectElement'
PASS tagJSWrapperPrototypeClass('object') is 'HTMLObjectElementPrototype'
PASS tagJSWrapperConstructorClass('object') is 'Function'
PASS tagJSWrapperClass('ol') is 'HTMLOListElement'
PASS tagJSWrapperPrototypeClass('ol') is 'HTMLOListElementPrototype'
PASS tagJSWrapperConstructorClass('ol') is 'Function'
PASS tagJSWrapperClass('optgroup') is 'HTMLOptGroupElement'
PASS tagJSWrapperPrototypeClass('optgroup') is 'HTMLOptGroupElementPrototype'
PASS tagJSWrapperConstructorClass('optgroup') is 'Function'
PASS tagJSWrapperClass('option') is 'HTMLOptionElement'
PASS tagJSWrapperPrototypeClass('option') is 'HTMLOptionElementPrototype'
PASS tagJSWrapperConstructorClass('option') is 'Function'
PASS tagJSWrapperClass('p') is 'HTMLParagraphElement'
PASS tagJSWrapperPrototypeClass('p') is 'HTMLParagraphElementPrototype'
PASS tagJSWrapperConstructorClass('p') is 'Function'
PASS tagJSWrapperClass('param') is 'HTMLParamElement'
PASS tagJSWrapperPrototypeClass('param') is 'HTMLParamElementPrototype'
PASS tagJSWrapperConstructorClass('param') is 'Function'
PASS tagJSWrapperClass('plaintext') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('plaintext') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('plaintext') is 'Function'
PASS tagJSWrapperClass('pre') is 'HTMLPreElement'
PASS tagJSWrapperPrototypeClass('pre') is 'HTMLPreElementPrototype'
PASS tagJSWrapperConstructorClass('pre') is 'Function'
PASS tagJSWrapperClass('q') is 'HTMLQuoteElement'
PASS tagJSWrapperPrototypeClass('q') is 'HTMLQuoteElementPrototype'
PASS tagJSWrapperConstructorClass('q') is 'Function'
PASS tagJSWrapperClass('rp') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('rp') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('rp') is 'Function'
PASS tagJSWrapperClass('rt') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('rt') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('rt') is 'Function'
PASS tagJSWrapperClass('ruby') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('ruby') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('ruby') is 'Function'
PASS tagJSWrapperClass('s') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('s') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('s') is 'Function'
PASS tagJSWrapperClass('samp') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('samp') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('samp') is 'Function'
PASS tagJSWrapperClass('script') is 'HTMLScriptElement'
PASS tagJSWrapperPrototypeClass('script') is 'HTMLScriptElementPrototype'
PASS tagJSWrapperConstructorClass('script') is 'Function'
PASS tagJSWrapperClass('section') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('section') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('section') is 'Function'
PASS tagJSWrapperClass('select') is 'HTMLSelectElement'
PASS tagJSWrapperPrototypeClass('select') is 'HTMLSelectElementPrototype'
PASS tagJSWrapperConstructorClass('select') is 'Function'
PASS tagJSWrapperClass('small') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('small') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('small') is 'Function'
PASS tagJSWrapperClass('span') is 'HTMLSpanElement'
PASS tagJSWrapperPrototypeClass('span') is 'HTMLSpanElementPrototype'
PASS tagJSWrapperConstructorClass('span') is 'Function'
PASS tagJSWrapperClass('strike') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('strike') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('strike') is 'Function'
PASS tagJSWrapperClass('strong') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('strong') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('strong') is 'Function'
PASS tagJSWrapperClass('style') is 'HTMLStyleElement'
PASS tagJSWrapperPrototypeClass('style') is 'HTMLStyleElementPrototype'
PASS tagJSWrapperConstructorClass('style') is 'Function'
PASS tagJSWrapperClass('sub') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('sub') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('sub') is 'Function'
PASS tagJSWrapperClass('sup') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('sup') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('sup') is 'Function'
PASS tagJSWrapperClass('table') is 'HTMLTableElement'
PASS tagJSWrapperPrototypeClass('table') is 'HTMLTableElementPrototype'
PASS tagJSWrapperConstructorClass('table') is 'Function'
PASS tagJSWrapperClass('tbody') is 'HTMLTableSectionElement'
PASS tagJSWrapperPrototypeClass('tbody') is 'HTMLTableSectionElementPrototype'
PASS tagJSWrapperConstructorClass('tbody') is 'Function'
PASS tagJSWrapperClass('td') is 'HTMLTableDataCellElement'
PASS tagJSWrapperPrototypeClass('td') is 'HTMLTableDataCellElementPrototype'
PASS tagJSWrapperConstructorClass('td') is 'Function'
PASS tagJSWrapperClass('textarea') is 'HTMLTextAreaElement'
PASS tagJSWrapperPrototypeClass('textarea') is 'HTMLTextAreaElementPrototype'
PASS tagJSWrapperConstructorClass('textarea') is 'Function'
PASS tagJSWrapperClass('tfoot') is 'HTMLTableSectionElement'
PASS tagJSWrapperPrototypeClass('tfoot') is 'HTMLTableSectionElementPrototype'
PASS tagJSWrapperConstructorClass('tfoot') is 'Function'
PASS tagJSWrapperClass('th') is 'HTMLTableHeaderCellElement'
PASS tagJSWrapperPrototypeClass('th') is 'HTMLTableHeaderCellElementPrototype'
PASS tagJSWrapperConstructorClass('th') is 'Function'
PASS tagJSWrapperClass('thead') is 'HTMLTableSectionElement'
PASS tagJSWrapperPrototypeClass('thead') is 'HTMLTableSectionElementPrototype'
PASS tagJSWrapperConstructorClass('thead') is 'Function'
PASS tagJSWrapperClass('title') is 'HTMLTitleElement'
PASS tagJSWrapperPrototypeClass('title') is 'HTMLTitleElementPrototype'
PASS tagJSWrapperConstructorClass('title') is 'Function'
PASS tagJSWrapperClass('tr') is 'HTMLTableRowElement'
PASS tagJSWrapperPrototypeClass('tr') is 'HTMLTableRowElementPrototype'
PASS tagJSWrapperConstructorClass('tr') is 'Function'
PASS tagJSWrapperClass('tt') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('tt') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('tt') is 'Function'
PASS tagJSWrapperClass('u') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('u') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('u') is 'Function'
PASS tagJSWrapperClass('ul') is 'HTMLUListElement'
PASS tagJSWrapperPrototypeClass('ul') is 'HTMLUListElementPrototype'
PASS tagJSWrapperConstructorClass('ul') is 'Function'
PASS tagJSWrapperClass('var') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('var') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('var') is 'Function'
PASS tagJSWrapperClass('wbr') is 'HTMLElement'
PASS tagJSWrapperPrototypeClass('wbr') is 'HTMLElementPrototype'
PASS tagJSWrapperConstructorClass('wbr') is 'Function'
PASS tagJSWrapperClass('xmp') is 'HTMLPreElement'
PASS tagJSWrapperPrototypeClass('xmp') is 'HTMLPreElementPrototype'
PASS tagJSWrapperConstructorClass('xmp') is 'Function'
